server/requirements.txt
Flask_Cors==3.0.7
pymongo==3.8.0
Flask_PyMongo==2.3.0
Flask==1.0.2
python_bcrypt==0.3.2
python-dotenv==0.10.2
